It seems like Obsidian is introducing `aliases`
https://forum.obsidian.md/t/obsidian-release-v0-9-16-insider-build/8795

I haven't paid for the catalyst payment profile, but it would suggest
this would be a solution for "unlinked mentions," so I'm going to use
this at this point in the import. I wonder if I really need to use the
`[]` list syntax vs the `- ` syntax.

If this works, then Obsidian is _it_ when it comes to the solution.

Tempted to throw $25 bucks at it so that I can test it...